+++ src/main/java/com/nirima/jenkins/plugins/docker/DockerTemplateBase.java
@@ -592,7 +592,7 @@ public class DockerTemplateBase implements Describable<DockerTemplateBase>, Seri
         }
 
         if (shmSize != null && shmSize > 0) {
-            final Long shmSizeInByte = Long.valueOf(shmSize * 1024L * 1024L);
+            final Long shmSizeInByte = shmSize * 1024L * 1024L;
             containerConfig.getHostConfig().withShmSize(shmSizeInByte);
         }
